+++ src/aqua/cli/diagnostic_config.py
@@ -228,12 +228,17 @@ diagnostic_config = {
     ],
     'ensemble': [
         {
-            'config_file': 'config_global_2D_ensemble.yaml',
+            'config_file': 'config_latlon_ensemble.yaml',
             'source_path': 'config/diagnostics/ensemble',
             'target_path': 'diagnostics/ensemble'
         },
         {
-            'config_file': 'config_timeseries_ensemble.yaml',
+            'config_file': 'config_single_model_timeseries_ensemble.yaml',
+            'source_path': 'config/diagnostics/ensemble',
+            'target_path': 'diagnostics/ensemble'
+        },
+        {
+            'config_file': 'config_multi_model_timeseries_ensemble.yaml',
             'source_path': 'config/diagnostics/ensemble',
             'target_path': 'diagnostics/ensemble'
         },