+++ src/aqua_diagnostics/__init__.py
@@ -6,7 +6,7 @@ from .global_biases import GlobalBiases, PlotGlobalBiases
 from .boxplots import Boxplots, PlotBoxplots
 from .ensemble import EnsembleTimeseries, EnsembleLatLon, EnsembleZonal
 from .ensemble import PlotEnsembleTimeseries, PlotEnsembleLatLon, PlotEnsembleZonal
-from .ensemble import retrieve_merge_ensemble_data
+from .ensemble import reader_retrieve_and_merge, merge_from_data_files, load_premerged_ensemble_dataset
 from .ecmean import PerformanceIndices, GlobalMean
 from .seaice import SeaIce, PlotSeaIce, Plot2DSeaIce
 
@@ -15,9 +15,11 @@ __all__ = ["NAO", "ENSO", "MJO",
            "Gregory", "SeasonalCycles", "Timeseries",
            "LatLonProfiles",
            "GlobalBiases", "PlotGlobalBiases",
+           "Radiation",
+           "EnsembleTimeseries", "EnsembleLatLon", 
+           "EnsembleZonal", "PlotEnsembleTimeseries", 
+           "PlotEnsembleLatLon", "PlotEnsembleZonal",
+           "reader_retrieve_and_merge", "merge_from_data_files", "load_premerged_ensemble_dataset",
            "Boxplots", "PlotBoxplots",
-           "EnsembleTimeseries", "EnsembleLatLon", "EnsembleZonal",
-           "PlotEnsembleTimeseries", "PlotEnsembleLatLon", "PlotEnsembleZonal",
-           "retrieve_merge_ensemble_data",
-           "PerformanceIndices", "GlobalMean",
-           "SeaIce", "PlotSeaIce", "Plot2DSeaIce"]
+           "GlobalMean", "PerformanceIndices", "SeaIce", "PlotSeaIce", "Plot2DSeaIce",
+           "Boxplots", "PlotBoxplots"]